+++ gdscript/src/main/kotlin/config/GdConfigFileType.kt
@@ -7,7 +7,7 @@ object GdConfigFileType : LanguageFileType(GdConfigLanguage) {
 
     override fun getName(): String = "GodotConfig file"
 
-    override fun getDescription(): String = "Godot's config file"
+    override fun getDescription(): String = "GodotConfig file"
 
     override fun getDefaultExtension(): String = "gdconf"