+++ gdscript/src/main/kotlin/config/psi/GdConfigTokenType.kt
@@ -1,0 +1,14 @@
+package config.psi
+
+import com.intellij.psi.tree.IElementType
+import gdscript.GdLanguage
+import org.jetbrains.annotations.NonNls
+
+class GdConfigTokenType(debugName: @NonNls String) :
+    IElementType(debugName, GdLanguage) {
+
+    override fun toString(): String {
+        return "GdConfigTokenType." + super.toString()
+    }
+
+}