+++ gdscript/src/main/kotlin/gdscript/library/GdLibraryManager.kt
@@ -17,9 +17,6 @@ import gdscript.utils.GdSdkUtil.sdkToVersion
 import gdscript.utils.GdSdkUtil.versionToSdkName
 import gdscript.utils.GdSdkUtil.versionToSdkUrl
 import gdscript.utils.GdSdkUtil.versionToSdkZip
-import kotlinx.coroutines.DelicateCoroutinesApi
-import kotlinx.coroutines.GlobalScope
-import kotlinx.coroutines.launch
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.decodeFromString
 import kotlinx.serialization.json.Json
@@ -33,115 +30,78 @@ import java.net.http.HttpResponse
 import java.nio.file.Files
 import java.nio.file.Paths
 import java.util.zip.ZipFile
-import javax.swing.JComboBox
+
 
 object GdLibraryManager {
 
-    var project: Project? = null
-
     val LIBRARY_NAME = "GdSdk"
 
-    @OptIn(DelicateCoroutinesApi::class)
-    fun listRemoteSdks(cb: JComboBox<String>) {
-        GlobalScope.launch {
-            val client = HttpClient.newBuilder().build()
-            val request = HttpRequest.newBuilder()
-                .uri(URI.create(SDKs_URL))
-                .build()
+    fun listAvailableSdks(): List<String> {
+        val client = HttpClient.newBuilder().build()
+        val request = HttpRequest.newBuilder()
+            .uri(URI.create(SDKs_URL))
+            .build()
 
-            val response = client.send(request, HttpResponse.BodyHandlers.ofString())
+        val response = client.send(request, HttpResponse.BodyHandlers.ofString())
 
-            // https://github.com/Kotlin/kotlinx.serialization/issues/993
-            @Suppress("PROVIDED_RUNTIME_TOO_LOW")
-            @Serializable
-            class GdSdk(val id: String, val name: String, val type: String, val path: String, val mode: String)
+        // https://github.com/Kotlin/kotlinx.serialization/issues/993
+        @Suppress("PROVIDED_RUNTIME_TOO_LOW")
+        @Serializable
+        class GdSdk(val id: String, val name: String, val type: String, val path: String, val mode: String)
 
-            Json.decodeFromString<Array<GdSdk>>(response.body())
-                .map { it.name.sdkToVersion() }
-                .sortedDescending()
-                .forEach { cb.addItem(it) }
-        }
+        return Json.decodeFromString<Array<GdSdk>>(response.body())
+            .map { it.name.sdkToVersion() }
+            .sortedDescending()
     }
 
-    fun listRegisteredSdks(): Iterable<Library> {
-        return ApplicationManager
-            .getApplication()
-            .getService(ModifiableModelsProvider::class.java)
-            .libraryTableModifiableModel.libraries.filter {
+    fun listRegisteredSdks(project: Project): Iterable<Library> {
+        return LibraryTablesRegistrar
+            .getInstance()
+            .getLibraryTable(project)
+            .modifiableModel
+            .libraries
+            .filter {
                 (it as LibraryEx).kind is GdLibraryKind
             }
     }
 
-    fun getLibrary(path: String): Library? {
-        return listRegisteredSdks().find {
-            ((it as LibraryEx).modifiableModel.properties as GdLibraryProperties).path == path
-        }
-    }
-
-    fun registerSdk(path: String): Library {
-        val name = path.replace('\\', '/').substringAfterLast("/")
-        val app = ApplicationManager.getApplication()
-        val modifier = LibraryTablesRegistrar.getInstance().libraryTable.modifiableModel
-        val library = modifier.createLibrary(name, GdLibraryKind)
-        val libModifier = library.modifiableModel as ModifiableModelEx
-        val props = libModifier.properties as GdLibraryProperties
-        props.path = path
-        props.version = name.sdkToVersion()
-        libModifier.properties = props
-        libModifier.addRoot("file://$path", OrderRootType.SOURCES)
-
-        app.invokeAndWait {
-            runWriteAction {
-                libModifier.commit()
-                modifier.commit()
-            }
-        }
-
-        return library
-    }
-
-    fun setUpLibrary(project: Project, path: String) {
-        if (path.isBlank()) return
-
-        getLibrary(path) ?: return
-
+    fun registerSdk(path: String, project: Project) {
         ApplicationManager.getApplication().invokeAndWait {
             runWriteAction {
-                var modifier = LibraryTablesRegistrar.getInstance().getLibraryTable(project).modifiableModel
-                modifier.getLibraryByName(LIBRARY_NAME)?.let { modifier.removeLibrary(it) }
-                modifier.commit()
+                val name = path.replace('\\', '/').substringAfterLast("/")
+                val modifier = LibraryTablesRegistrar.getInstance().getLibraryTable(project).modifiableModel
+                val library = modifier.createLibrary(name, GdLibraryKind)
 
-                modifier = LibraryTablesRegistrar.getInstance().getLibraryTable(project).modifiableModel
-                val library = modifier.createLibrary(LIBRARY_NAME)
-                val libraryModifier = library.modifiableModel
-                libraryModifier.addRoot("file://$path", OrderRootType.SOURCES)
-                libraryModifier.commit()
+                val libModifier = library.modifiableModel as ModifiableModelEx
+                val props = libModifier.properties as GdLibraryProperties
+                props.path = path
+                props.version = name.sdkToVersion()
+
+                libModifier.properties = props
+                libModifier.addRoot("file://$path", OrderRootType.SOURCES)
+
+                libModifier.commit()
                 modifier.commit()
 
                 val module = ModuleManager.getInstance(project).modules.first()
                 val rootModel = ModuleRootManager.getInstance(module).modifiableModel
-                rootModel.orderEntries.forEach {
-                    if (it is LibraryOrderEntry && it.libraryName == LIBRARY_NAME)
-                        rootModel.removeOrderEntry(it)
-                }
                 rootModel.addLibraryEntry(library)
                 rootModel.commit()
             }
         }
     }
 
-    fun clearSdks() {
-        if (project == null) return
+    fun clearSdks(project: Project) {
         ApplicationManager.getApplication().invokeAndWait {
             runWriteAction {
-                val modifier = LibraryTablesRegistrar.getInstance().getLibraryTable(project!!).modifiableModel
+                val modifier = LibraryTablesRegistrar.getInstance().getLibraryTable(project).modifiableModel
                 val tableModel = ApplicationManager
                     .getApplication()
                     .getService(ModifiableModelsProvider::class.java)
-                    .libraryTableModifiableModel
+                    .getLibraryTableModifiableModel(project)
 
                 modifier.getLibraryByName(LIBRARY_NAME)?.let { modifier.removeLibrary(it) }
-                listRegisteredSdks().forEach {
+                listRegisteredSdks(project).forEach {
                     it.rootProvider.getFiles(OrderRootType.SOURCES).forEach { dir ->
                         try {
                             FileUtils.deleteDirectory(File(dir.path))
@@ -174,7 +134,9 @@ object GdLibraryManager {
         val zipPath = Paths.get(zipName)
         val sdkName = version.versionToSdkName()
 
-        Files.createDirectory(Paths.get(directory, sdkName))
+        try {
+            Files.createDirectory(Paths.get(directory, sdkName))
+        } catch (e: Exception) {}
         val dir = Paths.get(directory, sdkName).toFile()
 
         URL(version.versionToSdkUrl())