+++ gdscript/src/main/kotlin/gdscript/settings/GdSettingsConfigurable.kt
@@ -2,7 +2,6 @@ package gdscript.settings
 
 import com.intellij.openapi.options.Configurable
 import com.intellij.openapi.project.Project
-import gdscript.library.GdLibraryManager
 import javax.swing.JComponent
 
 class GdSettingsConfigurable(val project: Project) : Configurable {
@@ -20,7 +19,6 @@ class GdSettingsConfigurable(val project: Project) : Configurable {
 
     override fun createComponent(): JComponent? {
         component = GdSettingsComponent(project)
-        GdLibraryManager.project = project
 
         return component?.panel
     }
@@ -29,7 +27,6 @@ class GdSettingsConfigurable(val project: Project) : Configurable {
         val settings = GdProjectSettingsState.getInstance(project).state
 
         return component?.hidePrivate != settings.hidePrivate
-            || component?.sdkPath != settings.sdkPath
             || component?.shortTyped != settings.shortTyped
             || component?.annotators != settings.annotators
             || component?.criticals != settings.criticals
@@ -45,18 +42,11 @@ class GdSettingsConfigurable(val project: Project) : Configurable {
         settings.criticals = component?.criticals ?: "ALERT,ATTENTION,CAUTION,CRITICAL,DANGER,SECURITY"
         settings.warnings = component?.warnings ?: "BUG,DEPRECATED,FIXME,HACK,TASK,TBD,TODO,WARNING"
         settings.notes = component?.notes ?: "INFO,NOTE,NOTICE,TEST,TESTING"
-
-        val oldSdk = settings.sdkPath
-        settings.sdkPath = component?.sdkPath
-
-        if (oldSdk == settings.sdkPath) return
-        GdLibraryManager.setUpLibrary(project, settings.sdkPath ?: "")
     }
 
     override fun reset() {
         val settings = GdProjectSettingsState.getInstance(project).state
         component?.hidePrivate = settings.hidePrivate
-        component?.sdkPath = settings.sdkPath
         component?.shortTyped = settings.shortTyped
         component?.annotators = settings.annotators
         component?.criticals = settings.criticals